requirements.txt
avro-python3==1.10.2
bitstring==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
grpcio==1.53.0
grpcio-tools==1.53.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
protobuf==4.22.1
pytest==7.3.0
requests==2.28.2
urllib3==1.26.15
